requirements.txt
googlemaps==4.10.0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
beautifulsoup4==4.12.2
lxml==4.9.3
